refactor(benchmark): separate detail route entries - Share the detail view across catalog and workspace pages - Point routes to dedicated page directories
